fix rpc sendRawTransaction to not push error to channel v1 sdk (#1018)
